// src/app/models/worker.ts
export interface Worker{
    Id: number
    JMBG: string
    FirstName: string
    LastName: string
    ParentName: string
    Address: string
    PhoneNumbers: string
    EmailAddresses: string
    Role: string
    HiredFromDate: Date
    HiredUntilDate: Date
    ExpoId: number
    IsExpoChief: boolean
}